Refactors and updates directory structure for workflow management
