+++ apps/web/messages/ar.json
@@ -1664,6 +1664,9 @@
                 "selectOption": "اختر خيارًا",
                 "directoryName": "اسم المجلد",
                 "generationPrompt": "موجه التوليد",
+                "modelOverride": "تجاوز نموذج الذكاء الاصطناعي",
+                "modelOverridePlaceholder": "مثل openai/gpt-4.1 أو openrouter/openai/gpt-4.1",
+                "modelOverrideHelperText": "اختياري. يستخدم النموذج المستورد أو السابق افتراضيًا إلا إذا قمت بتجاوزه هنا.",
                 "recommended": "مُوصى به",
                 "promptPlaceholder": "صف ما يجب توليده من عناصر لهذا المجلد...",
                 "promptHelperText": "كن محددًا بشأن نوع المحتوى الذي تريد توليده",
@@ -1805,6 +1808,28 @@
                 "syncDescription": "سحب التحديثات تلقائيًا من مستودع المصدر.",
                 "syncUpdateFailed": "فشل في تحديث إعدادات المزامنة",
                 "syncDisabled": "المزامنة معطلة",
+                "worksConfig": {
+                    "imported": "مستورد من works.yml",
+                    "applied": "مُطبق: {value}",
+                    "enabled": "مُفعّل",
+                    "disabled": "مُعطّل",
+                    "fields": {
+                        "sourceType": "نوع المصدر",
+                        "configName": "اسم التكوين",
+                        "initialPrompt": "الموجه الأولي",
+                        "model": "النموذج",
+                        "schedule": "الجدولة",
+                        "providers": "المزودون",
+                        "websiteRepo": "مستودع الموقع",
+                        "additionalAgents": "وكلاء إضافيون"
+                    },
+                    "sourceTypes": {
+                        "dataRepo": "مستودع البيانات",
+                        "awesomeReadme": "ملف README الرائع",
+                        "linkExisting": "مستودعات موجودة مرتبطة",
+                        "worksConfig": "works.yml"
+                    }
+                },
                 "generalSettings": "الإعدادات العامة",
                 "generalTitle": "الإعدادات العامة",
                 "directoryName": "اسم الدليل",
@@ -2348,15 +2373,19 @@
                 },
                 "detectedType": {
                     "dataRepo": "تم كشف مستودع بيانات",
-                    "awesomeReadme": "تم كشف قائمة Awesome"
+                    "awesomeReadme": "تم كشف قائمة Awesome",
+                    "worksConfig": "مستودع تكوين Works"
                 },
                 "badges": {
                     "dataRepo": "مستودع بيانات",
-                    "awesomeReadme": "قائمة Awesome"
+                    "awesomeReadme": "قائمة Awesome",
+                    "worksConfig": "works.yml"
                 },
                 "preview": {
                     "items": "عناصر",
-                    "categories": "فئات"
+                    "categories": "فئات",
+                    "schedule": "الجدولة: {cadence}",
+                    "providers": "المزودون: {providers}"
                 },
                 "nameLabel": "اسم الدليل",
                 "namePlaceholder": "مثلًا، دليلي الرائع",
@@ -2453,6 +2482,7 @@
                     "linkRepos": "ربط المستودعات"
                 },
                 "detectionFailed": "لم نتمكن من كشف الصيغة تلقائيًا. يرجى اختيار كيفية استيراد هذا المستودع.",
+                "detectionFailedMissingConfig": "لم يتم العثور على تكوين مدعوم. أضف config.yml أو works.yml أو works_config/works.yml إلى المستودع، أو اختر تنسيق استيراد يدوي أدناه.",
                 "errors": {
                     "urlRequired": "يرجى إدخال رابط مستودع GitHub",
                     "unsupportedFormat": "صيغة هذا المستودع غير مدعومة. يرجى استخدام صيغة مستودع بيانات أو قائمة Awesome.",
