+++ apps/web/messages/en.json
@@ -1808,6 +1808,28 @@
                 "syncDescription": "Automatically pull updates from the source repository.",
                 "syncUpdateFailed": "Failed to update sync settings",
                 "syncDisabled": "Sync Disabled",
+                "worksConfig": {
+                    "imported": "Imported from works.yml",
+                    "applied": "Applied: {value}",
+                    "enabled": "enabled",
+                    "disabled": "disabled",
+                    "fields": {
+                        "sourceType": "Source Type",
+                        "configName": "Config Name",
+                        "initialPrompt": "Initial Prompt",
+                        "model": "Model",
+                        "schedule": "Schedule",
+                        "providers": "Providers",
+                        "websiteRepo": "Website Repo",
+                        "additionalAgents": "Additional Agents"
+                    },
+                    "sourceTypes": {
+                        "dataRepo": "Data Repository",
+                        "awesomeReadme": "Awesome README",
+                        "linkExisting": "Linked Existing Repositories",
+                        "worksConfig": "works.yml"
+                    }
+                },
                 "generalSettings": "General Settings",
                 "generalTitle": "General Settings",
                 "directoryName": "Directory Name",
@@ -2351,15 +2373,19 @@
                 },
                 "detectedType": {
                     "dataRepo": "Data Repository Detected",
-                    "awesomeReadme": "Awesome List Detected"
+                    "awesomeReadme": "Awesome List Detected",
+                    "worksConfig": "Works config repository"
                 },
                 "badges": {
                     "dataRepo": "Data Repo",
-                    "awesomeReadme": "Awesome List"
+                    "awesomeReadme": "Awesome List",
+                    "worksConfig": "works.yml"
                 },
                 "preview": {
                     "items": "items",
-                    "categories": "categories"
+                    "categories": "categories",
+                    "schedule": "Schedule: {cadence}",
+                    "providers": "Providers: {providers}"
                 },
                 "nameLabel": "Directory Name",
                 "namePlaceholder": "e.g., My Awesome Directory",
