+++ apps/web/messages/nl.json
@@ -1664,6 +1664,9 @@
                 "selectOption": "Selecteer een optie",
                 "directoryName": "Mapnaam",
                 "generationPrompt": "Generatieprompt",
+                "modelOverride": "AI-model overschrijven",
+                "modelOverridePlaceholder": "bijv. openai/gpt-4.1 of openrouter/openai/gpt-4.1",
+                "modelOverrideHelperText": "Optioneel. Gebruikt standaard het geïmporteerde of vorige model, tenzij je het hier overschrijft.",
                 "recommended": "Aanbevolen",
                 "promptPlaceholder": "Beschrijf welke items voor deze map moeten worden gegenereerd...",
                 "promptHelperText": "Wees specifiek over het type content dat je wilt genereren",
@@ -1805,6 +1808,28 @@
                 "syncDescription": "Haal automatisch updates op uit de bronrepository.",
                 "syncUpdateFailed": "Kon synchronisatie-instellingen niet bijwerken",
                 "syncDisabled": "Synchronisatie uitgeschakeld",
+                "worksConfig": {
+                    "imported": "Geïmporteerd vanuit works.yml",
+                    "applied": "Toegepast: {value}",
+                    "enabled": "ingeschakeld",
+                    "disabled": "uitgeschakeld",
+                    "fields": {
+                        "sourceType": "Bronsoort",
+                        "configName": "Configuratienaam",
+                        "initialPrompt": "Initiële prompt",
+                        "model": "Model",
+                        "schedule": "Schema",
+                        "providers": "Providers",
+                        "websiteRepo": "Website-repository",
+                        "additionalAgents": "Extra agenten"
+                    },
+                    "sourceTypes": {
+                        "dataRepo": "Gegevensrepository",
+                        "awesomeReadme": "Awesome README",
+                        "linkExisting": "Gekoppelde bestaande repositories",
+                        "worksConfig": "works.yml"
+                    }
+                },
                 "generalSettings": "Algemene instellingen",
                 "generalTitle": "Algemene instellingen",
                 "directoryName": "Mapnaam",
@@ -2348,15 +2373,19 @@
                 },
                 "detectedType": {
                     "dataRepo": "Data-repository gedetecteerd",
-                    "awesomeReadme": "Awesome List gedetecteerd"
+                    "awesomeReadme": "Awesome List gedetecteerd",
+                    "worksConfig": "Works-configrepository"
                 },
                 "badges": {
                     "dataRepo": "Data Repo",
-                    "awesomeReadme": "Awesome List"
+                    "awesomeReadme": "Awesome List",
+                    "worksConfig": "works.yml"
                 },
                 "preview": {
                     "items": "items",
-                    "categories": "categorieën"
+                    "categories": "categorieën",
+                    "schedule": "Schema: {cadence}",
+                    "providers": "Providers: {providers}"
                 },
                 "nameLabel": "Directory-naam",
                 "namePlaceholder": "bijv. Mijn Awesome Directory",
@@ -2453,6 +2482,7 @@
                     "linkRepos": "Repositories koppelen"
                 },
                 "detectionFailed": "We konden het format niet automatisch detecteren. Selecteer hoe deze repository te importeren.",
+                "detectionFailedMissingConfig": "Geen ondersteunde configuratie gevonden. Voeg config.yml, works.yml of works_config/works.yml toe aan de repository, of kies een handmatig importformaat hieronder.",
                 "errors": {
                     "urlRequired": "Voer een GitHub-repository-URL in",
                     "unsupportedFormat": "Dit repository-format wordt niet ondersteund. Gebruik een Data-repository of Awesome List-format.",
