+++ packages/agent/src/import/import.module.ts
@@ -5,7 +5,7 @@ import { MarkdownGeneratorModule } from '../generators/markdown-generator/markdo
 import { WebsiteGeneratorModule } from '../generators/website-generator/website-generator.module';
 import { SourceRepoAnalyzerService } from './source-repo-analyzer.service';
 import { ImportExecutorService } from './import-executor.service';
-import { WorksConfigService } from './works-config.service';
+import { WorksConfigService } from '@src/works-config/services/works-config.service';
 
 @Module({
     imports: [FacadesModule, DataGeneratorModule, MarkdownGeneratorModule, WebsiteGeneratorModule],
