+++ packages/agent/src/import/index.ts
@@ -2,4 +2,3 @@ export * from './import.module';
 export * from './source-repo-analyzer.service';
 export * from './import-executor.service';
 export * from './enrichment-prompt.utils';
-export * from './works-config.service';
