+++ packages/agent/src/services/index.ts
@@ -14,7 +14,6 @@ export * from './directory-taxonomy.service';
 export * from './item-health.service';
 export * from './item-source-validation-scheduler.service';
 export * from './generator-form-schema.service';
-export * from './works-config-restore.service';
 export * from './utils/error-classification.utils';
 export * from './utils/error.utils';
 export * from './types/trigger-context.types';
