+++ packages/agent/src/works-config/index.ts
@@ -1,0 +1,5 @@
+export * from './services/works-config.service';
+export * from './services/works-config-restore.service';
+export * from './services/works-config-import-planner.service';
+export * from './services/works-config-import-applier.service';
+export * from './works-config-data';
